spike24/spike-ariadne-codegen/pyproject.toml
[tool.poetry]
name = "spike-ariadne-codegen"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.6"
httpx = "^0.28.1"
websockets = "^15.0.1"
python-dotenv = "1.1.0"

[tool.ariadne-codegen]
schema_path = "schema.graphql"
queries_path = "queries.graphql"
enable_custom_operations = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
